+++ CHANGELOG.md
@@ -14,9 +14,10 @@
 5. Drop ExaData from Database System Shapes because this is no longer the preferred implementation.
 6. Implement Version 2 of Cost Estimation / BoM ([Matches new online estimation tool](https://www.oracle.com/uk/cloud/costestimator.html))
     1. Block Storage Volume
-    1. File Storage System (Properties allow for usage Estimates)
+    1. File System (Properties allow for usage Estimates)
     1. Instance
     1. MySQL Database System
+    1. NoSQL Database
     1. Object Storage Bucket (Properties allow for usage Estimates)
 ### Bug Fixes
 1. In markdown generation switch to using documentation not definition